+++ shared_data/src/shared_data.cpp
@@ -13,8 +13,8 @@ SharedDataMutex<uint16_t> middle_IR_sensor(0);
 SharedDataMutex<uint16_t> right_IR_sensor(0);
 GyroReadingMutex gyro_data;
 // Encoder Data
-SharedDataMutex<int64_t> left_encoder_rotation_demand(0);
-SharedDataMutex<int64_t> right_encoder_rotation_demand(0);
+SharedDataMutex<int32_t> left_encoder_rotation_demand(0);
+SharedDataMutex<int32_t> right_encoder_rotation_demand(0);
 // Motor Control
 SharedDataMutex<bool> motor_action(false);
 SharedDataMutex<bool> calculate_corrections(false);
